Add test cases we want to support
I'm starting with test cases (without implementing the corresponding rules) to make sure I have covered a reasonable set of cases we'd like the linter to catch. I'm not normally a TDD-er but for this kind of thing it seems useful! Please let me know if there are common patterns not covered but the test cases that you'd like to see.

Test Plan:
These tests don't yet run (since the rules don't yet exist!).