Handle variables when setting or initializing state
Previously we only handled cases where the animated value was set to state directly (e.g. state = { color: new Animated.Value(0)}). Now we handle the case where a variable is set to an animated value, and that variable is then used to set state. We do this by iterating through every variable that we set to state, and checking if it was ever set to an Animated.Value instance (within the current scope)

Test Plan:
confirm a bunch more tests pass (test cases that involve setting a variable to Animated.Value(0) and then setting state with that variable)