Handle case where state is extracted to variable before torn down
We previously weren't handling the case where, instead of calling `this.state.foo.stopAnimation()`, we first extract the state as a variable and then call stopAnimation on that variable. Now we keep track of every variable we've extracted from the state and then check for matches when we teardown.

I also add another test case to make sure we handle the two most common ways of extracting the state into a variable.

Test Plan:
confirm we now properly detect teardowns even if the animated variable is first extracted from the state and then torn down (2 new test cases pass)