Support multiple components in a single file (#7)
* Support multiple components in a single file

Previously we stored the animations as a flat map keyed by the state field. This works for a single component but doesn't handle the case where there are multiple components in the same file, possibly with the same animated states. Now we store the animations keyed by a unique component identifier.

Test Plan:
confirm one additional test passes (the one asserting that both components must be torn down) (note: I had to tweak the test, previously it was passing in both cases because of the order of the teardowns)